+++ lib/dartactyl.dart
@@ -4,4 +4,4 @@
 library dartactyl;
 
 export 'src/client.dart';
-export 'src/models/models.dart';
+export 'models.dart';