+++ lib/src/models/egg.dart
@@ -1,6 +1,6 @@
 import 'package:json_annotation/json_annotation.dart';
 
-import 'models.dart';
+import '../../models.dart';
 
 part 'egg.g.dart';